Coding Hill's Algorithm
I finished the algoritm of Hill and I invite you to test it.
You could find some failures. In that case, please send me a mail
at yonnival0.8@gmail.com, and I will check my code to fix it.

GNPA
I've also wrote a small GNPA algoritm to remark the different
predictions we could to when there's not enough entropy

Enjoy !

Changes to be committed:
	new file:   ../GNPA/main.py
	new file:   ../GNPA/readme.md
	new file:   ../Hill's Algorithm/__pycache__/functions.cpython-311.pyc
	new file:   ../Hill's Algorithm/functions.py
	new file:   ../Hill's Algorithm/main.py
	new file:   ../Hill's Algorithm/readme.md
	modified:   __pycache__/functions.cpython-311.pyc
	modified:   functions.py
	modified:   main.py
	modified:   readme.md